+++ clevertap-xps/src/main/java/com/clevertap/android/xps/XiaomiPushProvider.java
@@ -13,12 +13,13 @@ import com.clevertap.android.sdk.CleverTapInstanceConfig;
 import com.clevertap.android.sdk.pushnotification.CTPushProvider;
 import com.clevertap.android.sdk.pushnotification.CTPushProviderListener;
 import com.clevertap.android.sdk.pushnotification.PushConstants;
+import com.clevertap.android.sdk.pushnotification.UnregistrableCTPushProvider;
 
 /**
  * Clevertap's Xiaomi Plugin Ref: {@link CTPushProvider}
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-public class XiaomiPushProvider implements CTPushProvider {
+public class XiaomiPushProvider implements CTPushProvider, UnregistrableCTPushProvider {
 
     private @NonNull
     final CTPushProviderListener ctPushListener;
@@ -29,8 +30,14 @@ public class XiaomiPushProvider implements CTPushProvider {
     @SuppressLint(value = "unused")
     public XiaomiPushProvider(@NonNull CTPushProviderListener ctPushListener, Context context,
             CleverTapInstanceConfig config) {
+        this(ctPushListener,context,config,true);
+    }
+
+    @SuppressLint(value = "unused")
+    public XiaomiPushProvider(@NonNull CTPushProviderListener ctPushListener, Context context,
+            CleverTapInstanceConfig config,Boolean isInit) {
         this.ctPushListener = ctPushListener;
-        this.miSdkHandler = new XiaomiSdkHandler(context, config);
+        this.miSdkHandler = new XiaomiSdkHandler(context, config,isInit);
     }
 
     @Override
@@ -62,6 +69,11 @@ public class XiaomiPushProvider implements CTPushProvider {
     @Override
     public void requestToken() {
         ctPushListener.onNewToken(miSdkHandler.onNewToken(), getPushType());
+    }
+
+    @Override
+    public void unregisterPush(final Context context) {
+        miSdkHandler.unregisterPush(context);
     }
 
     void setMiSdkHandler(@NonNull IMiSdkHandler sdkHandler) {